+++ plugins/org.preesm.codegen/src/org/preesm/codegen/printer/CodegenAbstractPrinter.java
@@ -213,7 +213,7 @@ public abstract class CodegenAbstractPrinter extends CodegenSwitch<CharSequence>
     return input.subSequence(lastLineBeginning, lastLineIndentationEnd).toString();
   }
 
-  PrinterState state = PrinterState.IDLE;
+  protected PrinterState state = PrinterState.IDLE;
 
   /**
    * Method used to change the current state of the printer.
@@ -241,26 +241,6 @@ public abstract class CodegenAbstractPrinter extends CodegenSwitch<CharSequence>
     // it. By default, use the first code block in the list.
     return 0;
   }
-
-  /**
-   * True if the visitor is currently printing {@link Variable} {@link CoreBlock#getDefinitions() definitions}
-   */
-  protected boolean printingDefinitions = false;
-
-  /**
-   * True if the visitor is currently printing {@link Variable} {@link CoreBlock#getDeclarations() declarations}
-   */
-  protected boolean printingDeclarations = false;
-
-  /**
-   * True if the visitor is currently printing the {@link CallBlock} {@link CoreBlock#getInitBlock() initBlock}
-   */
-  protected boolean printingInitBlock = false;
-
-  /**
-   * True if the visitor is currently printing the {@link LoopBlock} {@link CoreBlock#getInitBlock() initBlock}
-   */
-  protected boolean printingLoopBlock = false;
 
   /**
    * Reference to the {@link CoreBlock} currently printed
@@ -421,6 +401,8 @@ public abstract class CodegenAbstractPrinter extends CodegenSwitch<CharSequence>
     //
     // String concatenation is done manually because of complex
     // tabulation handling
+    setPrintedCoreBlock(coreBlock);
+
     StringConcatenation result = new StringConcatenation();
     final CharSequence coreBlockHeader = printCoreBlockHeader(coreBlock);
     result.append(coreBlockHeader);
@@ -451,6 +433,9 @@ public abstract class CodegenAbstractPrinter extends CodegenSwitch<CharSequence>
       result.newLineIfNotEmpty();
     }
     result.append(printCoreBlockFooter(coreBlock));
+
+    setPrintedCoreBlock(null);
+
     return result;
   }
 
@@ -489,6 +474,7 @@ public abstract class CodegenAbstractPrinter extends CodegenSwitch<CharSequence>
       result.append(indentationCoreBlock);
     }
     result.append(printDeclarationsFooter(declarations), indentationCoreBlock);
+    setState(PrinterState.IDLE);
     return result;
   }
 
@@ -513,6 +499,7 @@ public abstract class CodegenAbstractPrinter extends CodegenSwitch<CharSequence>
       result.append(indentationCoreBlock);
     }
     result.append(printCoreLoopBlockFooter(coreBlock.getLoopBlock()), indentationCoreBlock);
+    setState(PrinterState.IDLE);
     return result;
   }
 
@@ -537,6 +524,7 @@ public abstract class CodegenAbstractPrinter extends CodegenSwitch<CharSequence>
       result.append(indentationCoreBlock);
     }
     result.append(printCoreInitBlockFooter(coreBlock.getInitBlock()), indentationCoreBlock);
+    setState(PrinterState.IDLE);
     return result;
   }
 
@@ -567,6 +555,7 @@ public abstract class CodegenAbstractPrinter extends CodegenSwitch<CharSequence>
       result.append(indentationCoreBlock);
     }
     result.append(printDefinitionsFooter(definitions), indentationCoreBlock);
+    setState(PrinterState.IDLE);
     return result;
   }
 
@@ -657,7 +646,11 @@ public abstract class CodegenAbstractPrinter extends CodegenSwitch<CharSequence>
 
   @Override
   public CharSequence caseFunctionCall(final FunctionCall functionCall) {
-    return printFunctionCall(functionCall);
+    final StringBuilder res = new StringBuilder();
+    res.append(printPreFunctionCall(functionCall));
+    res.append(printFunctionCall(functionCall));
+    res.append(printPostFunctionCall(functionCall));
+    return res.toString();
   }
 
   @Override
@@ -1190,6 +1183,16 @@ public abstract class CodegenAbstractPrinter extends CodegenSwitch<CharSequence>
    * @return the printed {@link CharSequence}
    */
   public abstract CharSequence printFunctionCall(FunctionCall functionCall);
+
+  /**
+   * Method called right before {@link #printFunctionCall}
+   */
+  public abstract CharSequence printPreFunctionCall(FunctionCall functionCall);
+
+  /**
+   * Method called right after {@link #printFunctionCall}
+   */
+  public abstract CharSequence printPostFunctionCall(FunctionCall functionCall);
 
   /**
    * Method called to print a {@link PapifyFunctionCall}.