+++ plugins/org.preesm.model.pisdf/src/org/preesm/model/pisdf/util/FifoBreakingCycleDetector.java
@@ -2,6 +2,7 @@
  * Copyright or © or Copr. IETR/INSA - Rennes (2019) :
  *
  * Alexandre Honorat [alexandre.honorat@insa-rennes.fr] (2019)
+ * Antoine Morvan [antoine.morvan@insa-rennes.fr] (2019)
  *
  * This software is a computer program whose purpose is to help prototyping
  * parallel applications using dataflow formalism.